+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.1.2] - 2024-08-08
+
+### Fixed
+
+- Fixed issue with `protomoney.Sub()` that caused it to return results with
+  units/nanos components with mismatched signs (thanks [@sameaste92])
+
 ## [0.1.1] - 2022-02-08
 
 ### Added
@@ -26,6 +33,10 @@ The format is based on [Keep a Changelog], and this project adheres to
 [unreleased]: https://github.com/dogmatiq/dosh
 [0.1.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.1.0
 [0.1.1]: https://github.com/dogmatiq/dogma/releases/tag/v0.1.1
+
+<!-- contributors -->
+
+[@sameaste92]: https://github.com/sameaste92
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD